backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.22
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0
ibm-watsonx-ai==1.3.32
